feat: [김영재] 회원가입 페이지

+++ Spring/YoungjaeKim/demo/src/main/java/kr/eddi/demo/lectureClass/vue/problem/controller/form/RequestAccountIdForm.java
@@ -10,9 +10,9 @@ import lombok.ToString;
 @NoArgsConstructor
 public class RequestAccountIdForm {
 
-    private Long gameAccountId;
+    private int gameAccountId;
 
-    public RequestAccountIdForm(Long gameAccountId) {
+    public RequestAccountIdForm(int gameAccountId) {
         this.gameAccountId = gameAccountId;
     }
 }
